Refactor test cases for `parser`
